requirements.txt
python-telegram-bot[job-queue]
flask
requests
yfinance
python-dotenv
multitasking
pymongo
